requirements.txt
streamlit
transformers
streamlit
transformers
sentence-transformers
scikit-learn
fpdf
qrcode
pillow
python-dotenv